+++ src/utils/src/utils.h
@@ -4,7 +4,7 @@
 #include <string>
 
 std::string RandomString(uint8_t length);
-std::string RandomString(unsigned long seed, uint8_t length);
+std::string RandomString(uint32_t seed, uint8_t length);
 
 String TimeToStr(uint8_t hours, uint8_t minutes);
